Add W3C epubcheck tool to Guiguts
We already bundle the HTML and CSS validators. The epubcheck
tool validates epub files. It is therefore slightly different to the
other tools in that the user has to select a file to be checked,
rather than it checking the currently loaded file.

Packaging seems to work fine when generating a release.

+++ src/lib/Guiguts/ErrorCheck.pm
@@ -15,7 +15,7 @@ my $APOS = "\x{2019}";    # Curly apostrophe/right single quote
 
 # General error check window
 # Handles Bookloupe, Jeebies, HTML & CSS Validate, Tidy, Link Check
-# pphtml, pptxt, ppvimage, Spell Query and Load External Checkfile.
+# pphtml, pptxt, ppvimage, Spell Query, epubcheck and Load External Checkfile.
 sub errorcheckpop_up {
     my ( $textwindow, $top, $errorchecktype ) = @_;
     my ( $line, $lincol );
@@ -608,7 +608,9 @@ sub errorcheckrun {    # Runs error checks
     }
     $textwindow->focus;
     ::update_indicators();
-    return 1 if ::nofileloadedwarning();
+    unless ( $errorchecktype eq 'epubcheck' ) {    # Checks an epub, not the currently loaded file
+        return 1 if ::nofileloadedwarning();
+    }
 
     my $jartypes = [ [ 'JAR file', [ '.jar', ] ], [ 'All Files', ['*'] ], ];
     if ( $errorchecktype eq 'HTML Tidy' ) {
@@ -627,6 +629,11 @@ sub errorcheckrun {    # Runs error checks
                 \$::validatecsscommand, $jartypes );
             return 1 unless $::validatecsscommand;
         }
+    } elsif ( $errorchecktype eq 'epubcheck' ) {
+        unless ($::epubcheckcommand) {
+            ::locateExecutable( 'W3C epubcheck (epubcheck.jar)', \$::epubcheckcommand, $jartypes );
+            return 1 unless $::epubcheckcommand;
+        }
     }
     ::savesettings();
     $top->Busy( -recurse => 1 );
@@ -638,7 +645,21 @@ sub errorcheckrun {    # Runs error checks
       $errorchecktype eq 'Bookloupe'
       ? "^(/[$::allblocktypes](?=[\[\n]))|([$::allblocktypes]/(?=[\n]))"
       : "";
-    savetoerrortmpfile( $tmpfname, $striptext ) unless $errorchecktype eq 'Spell Query';
+
+    # For epubcheck, instead of saving current file as a temporary file,
+    # user chooses an epub file to check
+    if ( $errorchecktype eq 'epubcheck' ) {
+        my $types = [ [ 'Epub Files', ['.epub'] ], [ 'All Files', ['*'] ], ];
+        $tmpfname = $::lglobal{errorcheckpop}->getOpenFile(
+            -title      => 'File Name?',
+            -filetypes  => $types,
+            -initialdir => ::getsafelastpath()
+        );
+        return 1 unless $tmpfname;
+    } elsif ( $errorchecktype ne 'Spell Query' ) {    # No external tool, so no temp file needed
+        savetoerrortmpfile( $tmpfname, $striptext );
+    }
+
     if ( $errorchecktype eq 'HTML Tidy' ) {
         ::run( $::tidycommand, "-f", $errname, "-e", "-utf8", $tmpfname );
     } elsif ( $errorchecktype eq 'Nu HTML Check' or $errorchecktype eq "Nu XHTML Check" ) {
@@ -648,6 +669,9 @@ sub errorcheckrun {    # Runs error checks
         my $runner = ::runner::tofile( $errname, $errname );    # stdout & stderr
         $runner->run( "java", "-jar", $::validatecsscommand, "--profile=$::cssvalidationlevel",
             "file:$tmpfname" );
+    } elsif ( $errorchecktype eq 'epubcheck' ) {
+        my $runner = ::runner::tofile( $errname, $errname );    # stdout & stderr
+        $runner->run( "java", "-Xss2048k", "-jar", $::epubcheckcommand, $tmpfname );
     } elsif ( $errorchecktype eq 'pphtml' ) {
         ::run( "perl", "lib/ppvchecks/pphtml.pl", "-i", $tmpfname, "-o", $errname );
     } elsif ( $errorchecktype eq 'Link Check' ) {
@@ -669,7 +693,7 @@ sub errorcheckrun {    # Runs error checks
         spellqueryrun($errname);
     }
     $top->Unbusy;
-    unlink $tmpfname;
+    unlink $tmpfname unless $errorchecktype eq 'epubcheck';    # Don't delete the epub file
     return 0;
 }
